Avoid login when executing sharing UI tests
